requirements.txt
pyrebase==3.0.27
pyserial==3.4
